Update selectors after docutils update
We recently updated the packages we use for building Cilium's
documentation. Doing so, we switched to docutils v0.17, which introduces
some subtle changes in the generated HTML. In particular, the '<div
id="..." class="section">' nodes are replaced with '<section id="...">'
[0]. One unexpected consequence of this change is that the indexing for
the Algolia search now fails for the 'latest' branch, which received the
changes. This is because the selectors, in the JSON configuration file
in this repository, rely on the class name ".section" and not on the
"section" tag.

This commit update the selectors to use any of ".section" (class name)
or "section" (tag), so that indexing can work with all branches. We can
remove the ".section" in the future, when all supported branches rely on
Docutils v0.17+ (when releasing Cilium v1.14?).

[0] https://sourceforge.net/p/docutils/code/8472/

Signed-off-by: Quentin Monnet <quentin@isovalent.com>